+++ src/library/general/sleepwrappers.h
@@ -31,7 +31,7 @@ namespace libtas {
 /* Advance time when sleep call, depending on config and main thread.
  * Returns if the call was transfered.
  */
-bool transfer_sleep(const struct timespec &ts);
+void transfer_sleep(const struct timespec &ts, struct timespec *rem);
 
 /* Sleep USECONDS microseconds, or until a signal arrives that is not blocked
    or ignored.